Add files via upload
The objective of this project is to detect whether a person is with mask or without mask. In order to protect ourselves from the infections, polutions and deadful viruses, almost every one of us tend to wear a face mask. It becomes increasingly necessary to check if the people in the crowd wear face masks in most public gatherings such as Malls, Theaters, Parks, etc., People are forced by laws to wear face masks in public in many countries.These rules and laws were developed as an action to the exponential growth in cases and deaths in many areas. However, the process of monitoring large groups of people is becoming more difficult. The monitoring process involves the detection of anyone who is not wearing a face mask. Reports indicate that wearing face masks, while at work clearly reduces the risk of transmission. An efficient and economic approach of using AI to create a safe environment in a manufacturing setup, a hybrid model using Deep learning - Convolutional Neural Network and OpenCV for face mask detection has been proposed. The proposed model can be integrated with surveillance cameras to impede the COVID-19 transmission by allowing the detection of people who are wearing masks not wearing face masks.